+++ CHANGELOG.md
@@ -22,6 +22,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
 
 - Use global ratelimiters instead of local for each thread.
+- Prevent internal server error when IP address can't be extracted from a request.
 
 ## [5.0.0] - 2026-03-23